requirements/requirements.txt
# website
Django==3.1.2
djangorestframework==3.12.0
django-cors-headers==3.5.0